+++ src/backend/base/langflow/custom/custom_component/component.py
@@ -1,5 +1,5 @@
 import inspect
-from typing import Any, AsyncIterator, Callable, ClassVar, Generator, Iterator, List, Optional, Union
+from typing import Any, Callable, ClassVar, List, Optional, Union
 from uuid import UUID
 
 import yaml
@@ -13,26 +13,6 @@ from langflow.services.tracing.schema import Log
 from langflow.template.field.base import UNDEFINED, Output
 
 from .custom_component import CustomComponent
-
-
-def recursive_serialize_or_str(obj):
-    try:
-        if isinstance(obj, dict):
-            return {k: recursive_serialize_or_str(v) for k, v in obj.items()}
-        elif isinstance(obj, list):
-            return [recursive_serialize_or_str(v) for v in obj]
-        elif isinstance(obj, BaseModel):
-            return {k: recursive_serialize_or_str(v) for k, v in obj.model_dump().items()}
-        elif isinstance(obj, (AsyncIterator, Generator, Iterator)):
-            # contain memory addresses
-            # without consuming the iterator
-            # return list(obj) consumes the iterator
-            # return f"{obj}" this generates '<generator object BaseChatModel.stream at 0x33e9ec770>'
-            # it is not useful
-            return "Unconsumed Stream"
-        return str(obj)
-    except Exception:
-        return str(obj)
 
 
 class Component(CustomComponent):
