+++ src/backend/base/langflow/schema/schema.py
@@ -108,3 +108,35 @@ def build_output_logs(vertex, result) -> dict:
         outputs |= {name: OutputValue(message=message, type=_type).model_dump()}
 
     return outputs
+
+
+def recursive_serialize_or_str(obj):
+    try:
+        if isinstance(obj, dict):
+            return {k: recursive_serialize_or_str(v) for k, v in obj.items()}
+        elif isinstance(obj, list):
+            return [recursive_serialize_or_str(v) for v in obj]
+        elif isinstance(obj, BaseModel):
+            if hasattr(obj, "model_dump"):
+                obj_dict = obj.model_dump()
+            elif hasattr(obj, "dict"):
+                obj_dict = obj.dict()  # type: ignore
+            return {k: recursive_serialize_or_str(v) for k, v in obj_dict.items()}
+
+        elif isinstance(obj, (AsyncIterator, Generator, Iterator)):
+            # contain memory addresses
+            # without consuming the iterator
+            # return list(obj) consumes the iterator
+            # return f"{obj}" this generates '<generator object BaseChatModel.stream at 0x33e9ec770>'
+            # it is not useful
+            return "Unconsumed Stream"
+        elif hasattr(obj, "dict"):
+            return {k: recursive_serialize_or_str(v) for k, v in obj.dict().items()}
+        elif hasattr(obj, "model_dump"):
+            return {k: recursive_serialize_or_str(v) for k, v in obj.model_dump().items()}
+        elif issubclass(obj, BaseModel):
+            # This a type BaseModel and not an instance of it
+            return repr(obj)
+        return str(obj)
+    except Exception:
+        return str(obj)
